Actually do the refactor properly
